Move Apply and related classes to stream
