+++ mk/stream.py
@@ -126,6 +126,45 @@ class BindThunk(ThunkStream):
         return self.thunk().bind(self.goal)
 
 
+class Apply(Stream):
+    __slots__ = ('state', 'goal')
+
+    def __init__(self, state, goal):
+        self.state = state
+        self.goal = goal
+
+    def mplus(self, stream):
+        return Apply(self.state, MPlusGoal(stream, self.goal))
+
+    def bind(self, goal):
+        return Apply(self.state, BindGoal(self.goal, goal))
+
+    def next(self):
+        return None, self.goal(self.state)
+
+
+class MPlusGoal:
+    __slots__ = ('stream', 'goal')
+
+    def __init__(self, stream, goal):
+        self.stream = stream
+        self.goal = goal
+
+    def __call__(self, state):
+        return self.stream.mplus(self.goal(state))
+
+
+class BindGoal:
+    __slots__ = ('left', 'right')
+
+    def __init__(self, left, right):
+        self.left = left
+        self.right = right
+
+    def __call__(self, state):
+        return self.left(state).bind(self.right)
+
+
 def unfold(stream):
     while stream is not None:
         head, stream = stream.next()
